Early sockets implementation + test

+++ common/ISerializable.h
@@ -1,10 +1,13 @@
 #pragma once
+
+#include <cstdint>
+
 class ISerializable //interface
 {
 public:
 
 	//return the total size of the serialized object
-	virtual size_t serializeSize() const = 0;
+	virtual int32_t serializeSize() const = 0;
 
 	//return the char* corresponding to the serialized data of the calling object
 	virtual char*  serialize() = 0;
